+++ packages/furo-doc-helper/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.4](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/doc-helper@1.5.3...@furo/doc-helper@1.5.4) (2020-04-22)
+
+**Note:** Version bump only for package @furo/doc-helper
+
+
+
+
+
 ## [1.5.3](https://github.com/theNorstroem/FuroBaseComponents/compare/@furo/doc-helper@1.5.2...@furo/doc-helper@1.5.3) (2020-04-21)
 
 **Note:** Version bump only for package @furo/doc-helper